+++ src/logic/configure.ts
@@ -1,5 +1,6 @@
 import {
   assertProviderJson,
+  prepareSecurityValues,
   ProfileProviderDefaults,
   ProviderJson,
   SecurityScheme,
@@ -18,7 +19,6 @@ import { formatShellLog, LogCallback } from '../common/log';
 import { OutputStream } from '../common/output-stream';
 import { ProfileId } from '../common/profile';
 import { prepareEnvVariables } from '../templates/env';
-import { prepareSecurityValues } from './configure.utils';
 
 export async function updateEnv(
   provider: string,
@@ -70,12 +70,21 @@ export function handleProviderResponse(
 ): number {
   options?.logCb?.(`Installing provider: "${response.name}"`);
 
+  let parameters: { [key: string]: string } | undefined = undefined;
+  if (response.parameters) {
+    parameters = {};
+    for (const parameter of response.parameters) {
+      parameters[parameter.name] = parameter.default || '';
+    }
+  }
   const security = response.securitySchemes
-    ? prepareSecurityValues(response.name, response.securitySchemes, options)
+    ? prepareSecurityValues(response.name, response.securitySchemes)
     : [];
+
   // update super.json
   superJson.setProvider(response.name, {
     security,
+    parameters,
     file: options?.localProvider
       ? superJson.relativePath(options.localProvider)
       : undefined,
@@ -232,6 +241,24 @@ export async function installProvider(parameters: {
     }
   } else {
     parameters.options?.logCb?.(`No security schemes found to configure.`);
+  }
+  // inform user about configured parameters
+  if (providerInfo.parameters && providerInfo.parameters.length > 0) {
+    for (const parameter of providerInfo.parameters) {
+      let description = '';
+      if (parameter.description) {
+        description = ` with description "${parameter.description}"`;
+      }
+      if (parameter.default) {
+        parameters.options?.logCb?.(
+          `🆗 Parameter ${parameter.name}${description} configured with default value "${parameter.default}"`
+        );
+      } else {
+        parameters.options?.logCb?.(
+          `❌ Parameter ${parameter.name}${description} has not been configured.\nPlease, configure this parameter manualy in super.json on path: ${superJson.path}`
+        );
+      }
+    }
   }
 }